Fix and move code
